externalizando componente de listado de proyectos
