+++ src/app/@theme/components/index.ts
@@ -4,3 +4,4 @@ export * from './search-input/search-input.component';
 export * from './tiny-mce/tiny-mce.component';
 export * from './theme-settings/theme-settings.component';
 export * from './theme-switcher/theme-switcher.component';
+export * from './proyectos/proyectos.component';
