+++ src/app/@theme/components/proyectos/proyectos.component.scss
@@ -1,0 +1,36 @@
+
+.identado1 {
+    padding-left: 5%;
+}
+.identado2 {
+    padding-left: 10%;
+}
+.identado3 {
+    padding-left: 15%;
+}
+.identado4 {
+    padding-left: 20%;
+}
+.identado5 {
+    padding-left: 25%;
+}
+.identado6 {
+    padding-left: 30%;
+}
+.identado7 {
+    padding-left: 35%;
+}
+.identado8 {
+    padding-left: 35%;
+}
+.center {
+    margin: auto;
+    width: 50%;
+    padding: 10px;
+    text-align: center;
+}
+.activo {
+    padding: 1%;
+    color: #fff;
+    background-color:#8a7fff;
+}
