+++ src/app/pages/activos/gestion/gestion.component.html
@@ -1,7 +1,5 @@
+<ngx-proyectos [projectLocal]="project" (getProject)="getOrg($event)"></ngx-proyectos>
 <nb-card>
-    <nb-card-header>
-       Activos
-    </nb-card-header>
     <nb-card-body>
         <ng2-smart-table [settings]="settings" [source]="data"  (createConfirm)="onCreateConfirm($event)" (editConfirm)="onEditConfirm($event)">
         </ng2-smart-table>
