--- src/app/pages/configuracion/organization/proyectos/proyectos.component.html
@@ -1,23 +1,0 @@
-<nb-card>
-    <nb-card-header>
-        Proyectos
-    </nb-card-header>
-    <nb-card-body>
-        <div *ngFor="let project of listProjectAWS">
-            <div class="row identado{{level(project.id)}}">
-                <div *ngIf="project === projectLocal " class="row activo" >
-                    <a class="icon-container" (click)="asign(project)">
-                        <span><i class="nb-play"></i> {{project.nombre}}
-                        </span>
-                    </a>
-                </div>
-                <div *ngIf="project !== projectLocal " class="row" >
-                    <a class="icon-container" (click)="asign(project)" >
-                        <span><i class="nb-play"></i> {{project.nombre}}
-                        </span>
-                    </a>
-                </div>
-            </div>
-        </div>
-    </nb-card-body>
-</nb-card>
