--- src/app/pages/configuracion/organization/proyectos/proyectos.component.scss
@@ -1,141 +1,0 @@
-@import '../../../../@theme/styles/themes';
-@import '~bootstrap/scss/mixins/breakpoints';
-@import '~@nebular/theme/styles/global/bootstrap/breakpoints';
-@include nb-install-component() {
-    nb-card-body {
-        overflow: visible;
-    }
-    .input-group {
-        margin-bottom: 1rem;
-    }
-    .validation-checkboxes {
-        display: flex;
-        justify-content: space-between;
-        .custom-control {
-            margin-left: 1rem;
-        }
-    }
-    .demo-checkboxes {
-        display: flex;
-        flex-direction: column;
-        margin-bottom: 1rem;
-    }
-    .demo-radio {
-        display: flex;
-        flex-direction: column;
-        margin-bottom: 1rem;
-    }
-    .demo-disabled-checkbox-radio {
-        display: flex;
-        flex-direction: column;
-        margin-bottom: 1rem;
-    }
-    .demo-checkboxes-radio {
-        display: flex;
-        justify-content: space-between;
-    }
-    .demo-rating {
-        display: flex;
-        justify-content: space-between;
-        flex-wrap: wrap;
-    }
-    .star {
-        font-size: 1.5rem;
-        color: nb-theme(color-fg);
-    }
-    .filled {
-        color: nb-theme(color-fg);
-    }
-    // TODO: Replace with the card header styles mixin
-    .rating-header {
-        line-height: 2rem;
-        font-size: 1.25rem;
-        font-family: nb-theme(font-secondary);
-        font-weight: nb-theme(font-weight-bolder);
-        color: nb-theme(color-fg-heading);
-    }
-    .current-rate {
-        font-size: 1.5rem;
-        padding-left: 1rem;
-        color: nb-theme(color-fg-heading);
-    }
-    .full-name-inputs {
-        display: flex;
-    }
-    .input-group.has-person-icon {
-        position: relative;
-        .form-control {
-            padding-left: 3rem;
-        }
-        &::before {
-            content: '\F47D';
-            font-family: 'Ionicons';
-            font-size: 2rem;
-            position: absolute;
-            z-index: 100;
-            left: 1rem;
-            top: 0.25rem;
-        }
-    }
-    .dropdown {
-        min-width: 7rem;
-    }
-    .dropdown-menu {
-        width: auto;
-    }
-    .form-group label {
-        padding-right: 1rem;
-        text-align: right;
-    }
-    ngb-rating {
-        outline: none;
-    }
-    ngb-rating i {
-        color: nb-theme(color-success);
-        @include nb-for-theme(cosmic) {
-            color: nb-theme(color-primary);
-        }
-    }
-    @include media-breakpoint-down(xs) {
-        button:not(.btn-icon) {
-            padding: 0.75rem 1rem;
-            font-size: 0.75rem;
-        }
-    }
-    .identado1 {
-        padding-left: 5%;
-    }
-    .identado2 {
-        padding-left: 10%;
-    }
-    .identado3 {
-        padding-left: 15%;
-    }
-    .identado4 {
-        padding-left: 20%;
-    }
-    .identado5 {
-        padding-left: 25%;
-    }
-    .identado6 {
-        padding-left: 30%;
-    }
-    .identado7 {
-        padding-left: 35%;
-    }
-    .identado8 {
-        padding-left: 35%;
-    }
-    .center {
-        margin: auto;
-        width: 50%;
-        padding: 10px;
-        text-align: center;
-    }
-    .activo {
-        padding: 1%;
-        color: #fff;
-        background-color:#8a7fff;
-    }
-
-}
